# openapi.yaml
openapi: 3.1.0
info:
  title: Eigora API
  description: Physics simulation backend for the Eigora platform.
  version: 0.1.0
paths:
  /qm/eigenstates:
    post:
      tags:
      - Quantum Mechanics
      summary: Eigenstates
      description: 'Solve the time-independent Schrödinger equation and return eigenstates.


        Accepts either a named potential (Option A) or a custom V(x) array (Option
        B).'
      operationId: eigenstates_qm_eigenstates_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EigenstatesRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EigenstatesResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /qm/separable-state:
    post:
      tags:
      - Quantum Mechanics
      summary: Separable State
      description: 'One eigenstate of a system that separates along x and y.


        The potential is given per axis, so `infinite_well` on both axes is a

        particle in a 2D box, `harmonic` on both is a 2D trap, and the axes may

        differ. Each axis is solved analytically when its potential has a known

        solution and numerically otherwise; the state is the product of the two.'
      operationId: separable_state_qm_separable_state_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeparableStateRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeparableStateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /qm/single-atom-state:
    post:
      tags:
      - Quantum Mechanics
      summary: Single Atom State
      description: 'Compute the single-atom state for a given potential and quantum
        numbers.


        Accepts either a named potential (Option A) or a custom V(x) array (Option
        B).'
      operationId: single_atom_state_qm_single_atom_state_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SingleAtomStateRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleAtomStateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    BarrierParams:
      properties:
        height:
          type: number
          exclusiveMinimum: 0.0
          title: Height
          default: 2.0
        width:
          type: number
          exclusiveMinimum: 0.0
          title: Width
          default: 1.0
        x0:
          type: number
          title: X0
          default: 0.0
      type: object
      title: BarrierParams
    DoubleWellParams:
      properties:
        a:
          type: number
          exclusiveMinimum: 0.0
          title: A
          default: 1.0
        b:
          type: number
          exclusiveMinimum: 0.0
          title: B
          default: 4.0
      type: object
      title: DoubleWellParams
    EigenstatesRequest:
      properties:
        grid:
          $ref: '#/components/schemas/GridSchema'
        potential:
          $ref: '#/components/schemas/PotentialSchema'
        n_states:
          type: integer
          maximum: 20.0
          minimum: 1.0
          title: N States
          default: 6
      type: object
      required:
      - potential
      title: EigenstatesRequest
    EigenstatesResponse:
      properties:
        x:
          items:
            type: number
          type: array
          title: X
        potential:
          items:
            type: number
          type: array
          title: Potential
        energies:
          items:
            type: number
          type: array
          title: Energies
        wavefunctions:
          items:
            items:
              type: number
            type: array
          type: array
          title: Wavefunctions
        n_states:
          type: integer
          title: N States
      type: object
      required:
      - x
      - potential
      - energies
      - wavefunctions
      - n_states
      title: EigenstatesResponse
    FiniteWellParams:
      properties:
        depth:
          type: number
          exclusiveMinimum: 0.0
          title: Depth
          default: 5.0
        width:
          type: number
          exclusiveMinimum: 0.0
          title: Width
          default: 4.0
        x0:
          type: number
          title: X0
          default: 0.0
      type: object
      title: FiniteWellParams
    Grid2DSchema:
      properties:
        x_min:
          type: number
          title: X Min
          default: -10.0
        x_max:
          type: number
          title: X Max
          default: 10.0
        y_min:
          type: number
          title: Y Min
          default: -10.0
        y_max:
          type: number
          title: Y Max
          default: 10.0
        nx:
          type: integer
          maximum: 2048.0
          minimum: 64.0
          title: Nx
          default: 512
        ny:
          type: integer
          maximum: 2048.0
          minimum: 64.0
          title: Ny
          default: 512
      type: object
      title: Grid2DSchema
    Grid3DSchema:
      properties:
        x_min:
          type: number
          title: X Min
          default: -10.0
        x_max:
          type: number
          title: X Max
          default: 10.0
        y_min:
          type: number
          title: Y Min
          default: -10.0
        y_max:
          type: number
          title: Y Max
          default: 10.0
        z_min:
          type: number
          title: Z Min
          default: -10.0
        z_max:
          type: number
          title: Z Max
          default: 10.0
        nx:
          type: integer
          maximum: 512.0
          minimum: 32.0
          title: Nx
          default: 128
        ny:
          type: integer
          maximum: 512.0
          minimum: 32.0
          title: Ny
          default: 128
        nz:
          type: integer
          maximum: 512.0
          minimum: 32.0
          title: Nz
          default: 128
      type: object
      title: Grid3DSchema
    GridSchema:
      properties:
        x_min:
          type: number
          title: X Min
          default: -10.0
        x_max:
          type: number
          title: X Max
          default: 10.0
        n_points:
          type: integer
          maximum: 2048.0
          minimum: 64.0
          title: N Points
          default: 512
      type: object
      title: GridSchema
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HarmonicParams:
      properties:
        omega:
          type: number
          exclusiveMinimum: 0.0
          title: Omega
          default: 1.0
        x0:
          type: number
          title: X0
          default: 0.0
      type: object
      title: HarmonicParams
    InfiniteWellParams:
      properties:
        width:
          type: number
          exclusiveMinimum: 0.0
          title: Width
          default: 4.0
        x0:
          type: number
          title: X0
          default: 0.0
      type: object
      title: InfiniteWellParams
    LobeMeshSchema:
      properties:
        positions:
          type: string
          title: Positions
        normals:
          type: string
          title: Normals
        indices:
          type: string
          title: Indices
        vertex_count:
          type: integer
          title: Vertex Count
        triangle_count:
          type: integer
          title: Triangle Count
      type: object
      required:
      - positions
      - normals
      - indices
      - vertex_count
      - triangle_count
      title: LobeMeshSchema
      description: 'One signed isosurface shell as a triangle mesh, packed for direct
        upload to

        a WebGL BufferGeometry. Each string is base64 of the little-endian raw

        bytes: positions/normals are flat float32 [x,y,z,...]; indices are flat

        uint32 triangle vertex indices.'
    PotentialSchema:
      properties:
        type:
          $ref: '#/components/schemas/PotentialType'
        params:
          anyOf:
          - $ref: '#/components/schemas/HarmonicParams'
          - $ref: '#/components/schemas/InfiniteWellParams'
          - $ref: '#/components/schemas/FiniteWellParams'
          - $ref: '#/components/schemas/BarrierParams'
          - $ref: '#/components/schemas/StepParams'
          - $ref: '#/components/schemas/DoubleWellParams'
          - type: 'null'
          title: Params
        values:
          anyOf:
          - items:
              type: number
            type: array
          - type: 'null'
          title: Values
          description: Custom V(x) values, one per grid point (Option B)
      type: object
      required:
      - type
      title: PotentialSchema
    PotentialType:
      type: string
      enum:
      - harmonic
      - infinite_well
      - finite_well
      - barrier
      - step
      - double_well
      - free
      - custom
      title: PotentialType
    SeparableStateRequest:
      properties:
        grid:
          $ref: '#/components/schemas/Grid2DSchema'
        potential_x:
          $ref: '#/components/schemas/PotentialSchema'
        potential_y:
          $ref: '#/components/schemas/PotentialSchema'
        n1:
          type: integer
          maximum: 99.0
          minimum: 0.0
          title: N1
          description: Index of the state along x, 0 being the ground state. This
            is a position in the ascending list of states, not the physical quantum
            number -- those differ per potential (a box starts at n=1, an oscillator
            at n=0) and are returned in `label`.
          default: 0
        n2:
          type: integer
          maximum: 99.0
          minimum: 0.0
          title: N2
          description: Index of the state along y.
          default: 0
        n_states:
          type: integer
          maximum: 100.0
          minimum: 1.0
          title: N States
          description: States computed per axis that has no analytic solution.
          default: 20
      type: object
      required:
      - potential_x
      - potential_y
      title: SeparableStateRequest
      description: 'One eigenstate of a system that separates along x and y.


        The potential is given per axis, so a 2D well is `infinite_well` on both

        axes, an anisotropic trap is `harmonic` on both with different omega, and

        the two axes may differ freely.'
    SeparableStateResponse:
      properties:
        x:
          items:
            type: number
          type: array
          title: X
        y:
          items:
            type: number
          type: array
          title: Y
        psi_x:
          items:
            type: number
          type: array
          title: Psi X
          description: Normalised factor along x
        psi_y:
          items:
            type: number
          type: array
          title: Psi Y
          description: Normalised factor along y
        potential_x:
          items:
            type: number
          type: array
          title: Potential X
          description: V(x) sampled on the x axis
        potential_y:
          items:
            type: number
          type: array
          title: Potential Y
          description: V(y) sampled on the y axis
        energy:
          type: number
          title: Energy
        energy_x:
          type: number
          title: Energy X
        energy_y:
          type: number
          title: Energy Y
        label:
          items:
            type: integer
          type: array
          title: Label
          description: Physical quantum numbers of the state
        quantum_numbers:
          items:
            type: string
          type: array
          title: Quantum Numbers
          description: Names of those quantum numbers
        degeneracy:
          type: integer
          title: Degeneracy
          description: How many states share this energy
        is_exact:
          type: boolean
          title: Is Exact
          description: False if either axis had to be solved numerically
      type: object
      required:
      - x
      - y
      - psi_x
      - psi_y
      - potential_x
      - potential_y
      - energy
      - energy_x
      - energy_y
      - label
      - quantum_numbers
      - degeneracy
      - is_exact
      title: SeparableStateResponse
      description: 'The state as its two one-dimensional factors: psi(x, y) = psi_x(x)
        * psi_y(y).


        Sending the factors rather than the (nx, ny) field keeps the payload

        proportional to nx + ny instead of nx * ny; the client reconstructs the

        field with an outer product.'
    SingleAtomStateRequest:
      properties:
        grid:
          $ref: '#/components/schemas/Grid3DSchema'
        Z:
          type: integer
          maximum: 100.0
          minimum: 1.0
          title: Z
          description: Atomic number
          default: 1
        n:
          type: integer
          maximum: 10.0
          minimum: 1.0
          title: N
          description: Principal quantum number
          default: 1
        l:
          type: integer
          maximum: 9.0
          minimum: 0.0
          title: L
          description: Orbital angular momentum quantum number
          default: 0
        m:
          type: integer
          title: M
          description: Magnetic quantum number
          default: 0
      type: object
      title: SingleAtomStateRequest
    SingleAtomStateResponse:
      properties:
        positive:
          anyOf:
          - $ref: '#/components/schemas/LobeMeshSchema'
          - type: 'null'
        negative:
          anyOf:
          - $ref: '#/components/schemas/LobeMeshSchema'
          - type: 'null'
        bound_radius:
          type: number
          title: Bound Radius
        Z:
          type: integer
          title: Z
        n:
          type: integer
          title: N
        l:
          type: integer
          title: L
        m:
          type: integer
          title: M
      type: object
      required:
      - positive
      - negative
      - bound_radius
      - Z
      - n
      - l
      - m
      title: SingleAtomStateResponse
    StepParams:
      properties:
        height:
          type: number
          title: Height
          default: 1.0
        x0:
          type: number
          title: X0
          default: 0.0
      type: object
      title: StepParams
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError